Fix material fill light level access crash

+++ src/Engine/rdMaterial.c
@@ -449,7 +449,7 @@ int rdMaterial_GetFillColor(rdVector3* pOutColor, rdMaterial* pMaterial, rdColor
 			if (pColormap)
 			{
 				if(lightLevel >= 0)
-					paletteIndex = ((int*)pColormap->lightlevel)[paletteIndex * 64 + lightLevel];
+					paletteIndex = ((uint8_t*)pColormap->lightlevel)[paletteIndex * 64 + lightLevel];
 
 				pOutColor->x = (float)pColormap->colors[paletteIndex].r / 255.0f;
 				pOutColor->y = (float)pColormap->colors[paletteIndex].g / 255.0f;
